backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
marshmallow==3.21.1
python-dotenv==1.0.1
pytest==8.3.3
